Draw naive dotted lines manually
First issues noticed:
- the rounding in this example and in the library are different
- on thin 45° lines, due to the distance between dots being too small,
dots are not always separated by gaps
- almost horizontal / vertical dotted lines look wonky
- on 45° lines, dots can look too big or too small

The last 3 problems are especially noticeable on thin lines =>
why not make a specific code for thin lines like for the rectangle?
This code could keep half the points in the iterator instead of trying
to draw small dots which look like rectangles anyway.